added further tests for address prefix rule
